refactor: find techs by tag into service
